requirements.txt
tensorflow==2.15.0
keras==2.15.0
pillow
numpy
streamlit
